move vtt functions from app to lib/vtt.py and make toml config
